Fix for #1950: make delete_zone an API request. next: only offer the option from org_management.
