+++ kalite/control_panel/api_urls.py
@@ -1,0 +1,5 @@
+from django.conf.urls.defaults import patterns, url
+
+urlpatterns = patterns(__package__ + '.api_views',
+    url(r'zone/(?P<zone_id>\w+)/delete$', 'delete_zone', {}, 'delete_zone'),
+)
