+++ kalite/control_panel/static/js/control_panel/zone_management.js
@@ -3,9 +3,11 @@ $(function () {
         force_sync();
     })
 
+    /*
     $(".zone-delete-link").click(function() {
         if (confirm(sprintf(gettext("Are you sure you want to delete sharing network '%(network_name)s'?"), {network_name: event.target.getAttribute("value")}))) {
-            window.location.href = DELETE_ZONE_URL;
+            doRequest(DELETE_ZONE_URL);
         }
     });
+    */
 })
