+++ kalite/control_panel/templates/control_panel/zone_management.html
@@ -55,17 +55,6 @@
                     <a class="green_button" href="#">{% trans "Sync via USB" %}</a>
                 </li-->
             {% endblock download_button %}
-            <li>
-                <ul id="zone-management-options">
-                        <li>
-                            <a href="#" class="zone-delete-link btn btn-danger" title="{% trans 'Delete this sharing network' %}" value="{{ zone.name }}">
-                                {% trans 'Delete this sharing network' %}<i class="icon-trash"></i>
-                            </a>
-                        </li>
-                    {% if not facilities and not devices %}{# can only happen on the central server #}
-                    {% endif %}
-                </ul>
-            </li>
 {% endblock buttons %}
 
 {% block control_panel_content %}
