+++ kalite/control_panel/urls.py
@@ -1,11 +1,11 @@
-from django.conf.urls.defaults import patterns, url
+from django.conf.urls.defaults import patterns, url, include
 
+from . import api_urls
 
 urlpatterns = patterns(__package__ + '.views',
     # Zone
     url(r'zone/(?P<zone_id>\w+)/$', 'zone_management', {}, 'zone_management'),
     url(r'zone/(?P<zone_id>\w+)/edit$', 'zone_form', {}, 'zone_form'),
-    url(r'zone/(?P<zone_id>\w+)/delete$', 'delete_zone', {}, 'delete_zone'),
 
     # Device
     url(r'zone/(?P<zone_id>\w+)/device/(?P<device_id>\w+)/$', 'device_management', {}, 'device_management'),
@@ -16,4 +16,6 @@ urlpatterns = patterns(__package__ + '.views',
     url(r'zone/(?P<zone_id>\w+)/facility/(?P<facility_id>\w+)/management/group/(?P<group_id>\w+)/$', 'facility_management', {}, 'facility_management'),
 
     url(r'account/$', 'account_management', {}, 'account_management'),
+
+    url(r'^api/', include(api_urls)),
 )
